Renamed some more files for the submission flow reorg.
Files checked in: DataRepo/management/commands/profile_submission_page.py DataRepo/tests/views/upload/test_validation.py DataRepo/views/upload/__init__.py DataRepo/views/upload/validation.py DataRepo/tests/views/upload/test_submission.py DataRepo/views/upload/submission.py
