+++ DataRepo/views/upload/__init__.py
@@ -1,5 +1,5 @@
+from .submission import DataValidationView, validation_disabled
 from .upload import upload
-from .validation import DataValidationView, validation_disabled
 
 __all__ = [
     "upload",
